Basic keyboard-controlled timer. The timer can be started with `space` and stopped with `p`. Pressing `space` again restarts the timer (from 0).
This should work well as the global timer. It currently updates every 100ms, but can be bumped to as low as 4ms (per the HTML5 spec for timer clamping). Performance wise, anything less than 20ms uses a large amount of CPU time, so 100 should be used as the default unless otherwise configured. This shouldn't be an issue, though, as the timer does not actual count elapsed time, and instead references the current wall-clock time and calculates the difference from the time that the timer started. This should allow individual timers to attach to and detach from the global timer without a concern of drift.