Add basic segmenting This commit introduces the concept of Segments and Splits, where Splits are just a collection of Segments associated with a single timer. A Splits instance is responsible for tracking the currently-active Segment and updating it when the timer ticks.
I'm not really a fan of the current dependency model, where the Timer gets injected into each Segment, but timer events have to be proxied through the Splits instance. It would be nice to have an Event bus for various events that any object can subscribe to. This will also help with later componentization, as third-party components will have a simple interface to the event system without relying on instance sharing.