Proper components via HTML5 imports As always, HTML5 has a wonderful feature that allows loading and importing other documents into another via a `<link rel="import" ...>` element. This, combined with a tiny bit of javascript to create those elements dynamically (see `./lib/component_loader.js`), means components can be arbitrarily loaded into the DOM at any time as if they existed there from the beginning.
With that, all of the previously-defined components have been moved to a `components` folder. Components are only required to have a `template.html` file that registers the component with rivets (`rivets.component['my-component'] = {...}`). Everything else, including the markup, bindings, and styles of the component are left as implementation details, so long as everything is imported into the `template.html`. See `./components/timer/*` for a good example of this.